refactor: move test-only exports to dedicated testing.go file
NewNodeNetworkConfigReconcilerForTesting and RestoreOnReconcileFailure are needed only by test code in dependent packages. Moving them to testing.go (a non-_test.go file so they remain importable from other packages' _test.go files) keeps the production nodenetworkconfig.go free of test-only exports. GetNodeNetworkConfig was not used anywhere and has been dropped entirely.
