+++ pkg/reconciler/common/nodenetworkconfig.go
@@ -81,37 +81,6 @@ type NodeNetworkConfigReconciler struct {
 	localASN                  int64
 }
 
-// NewNodeNetworkConfigReconcilerForTesting creates a NodeNetworkConfigReconciler
-// with a caller-supplied HealthCheckerInterface. This bypasses the real
-// healthcheck initialisation (LoadConfig, TCP dialer) and is intended for
-// unit tests that need to inject a stub or mock healthchecker.
-// Do NOT use this in production code.
-func NewNodeNetworkConfigReconcilerForTesting(
-	clusterClient client.Client,
-	logger logr.Logger,
-	configApplier ConfigApplier,
-	nodeNetworkConfigPath string,
-	opts ReconcilerOptions,
-	hc healthcheck.HealthCheckerInterface,
-) (*NodeNetworkConfigReconciler, error) {
-	reconciler := &NodeNetworkConfigReconciler{
-		client:                    clusterClient,
-		logger:                    logger,
-		configApplier:             configApplier,
-		healthChecker:             hc,
-		NodeNetworkConfigPath:     nodeNetworkConfigPath,
-		restoreOnReconcileFailure: opts.RestoreOnReconcileFailure,
-	}
-
-	var err error
-	reconciler.NodeNetworkConfig, err = ReadNodeNetworkConfig(reconciler.NodeNetworkConfigPath)
-	if err != nil && !errors.Is(err, os.ErrNotExist) {
-		return nil, fmt.Errorf("error reading NodeNetworkConfig from disk: %w", err)
-	}
-
-	return reconciler, nil
-}
-
 // NewNodeNetworkConfigReconciler creates a new NodeNetworkConfigReconciler.
 func NewNodeNetworkConfigReconciler(
 	clusterClient client.Client,
@@ -451,15 +420,4 @@ func SetStatusWithError(
 	}
 
 	return nil
-}
-
-// GetNodeNetworkConfig returns the current in-memory NodeNetworkConfig.
-func (r *NodeNetworkConfigReconciler) GetNodeNetworkConfig() *v1alpha1.NodeNetworkConfig {
-	return r.NodeNetworkConfig
-}
-
-// RestoreOnReconcileFailure returns whether the reconciler restores the previous
-// configuration when reconciliation fails.
-func (r *NodeNetworkConfigReconciler) RestoreOnReconcileFailure() bool {
-	return r.restoreOnReconcileFailure
 }
